+++ requirements.txt
@@ -13,4 +13,3 @@ pyopenssl>=17.3.0
 pyqrllib>=0.2.9
 six>=1.9
 click==6.7
-pycrypto
