+++ test-requirements.txt
@@ -7,3 +7,4 @@ mock>=2.0.0
 flake8
 timeout-decorator
 nose
+pycrypto
